+++ test/snapshot/core/plain_nullable_primitive_params/intended/main.go
@@ -24,7 +24,7 @@ func floatBranch(value any) float64 {
 
 func intBranch(value any) int {
 	if value != nil {
-		return int(int32((hxrt.Int32Wrap(value.(int)) + hxrt.Int32Wrap(1))))
+		return int((hxrt.Int32Wrap(value.(int)) + hxrt.Int32Wrap(1)))
 	}
 	return -1
 }
